+++ R/col_plan.R
@@ -150,12 +150,11 @@ check_span_structure_dots <- function(x) {
     x_names <- names(x)
 
     if (is.null(x_names) || !all(nzchar(x_names))) {
-        abort(
-            paste0(
-                "Entries of a span_stucture must be named:\n ",
-                format(caller_call())
-            ),
-            call = caller_call()
+        span_call <- caller_call()
+
+        cli::cli_abort(
+            "Entries of a span_stucture must be named:\n {format(span_call)}",
+            call = span_call
         )
     }
 
@@ -174,23 +173,15 @@ check_span_structure_dots <- function(x) {
                     } else if (is_valid_quo_call(x)) {
                         return(eval_tidy(x))
                     } else {
-                        abort(
-                            message = paste0(
-                                "Invalid entry: `",
-                                format(x),
-                                "`\n",
-                                "Only selection helpers (See <https://tidyselect.r-lib.org/reference>), ",
-                                " or unquoted expressions representing variable names ",
-                                " can be entered as contents.",
-                                " Changing the names of individual variables using new_name = old_name syntax is allowable"
-                            ),
+                        cli::cli_abort(
+                            message = "Invalid entry: `{format(x)}`\nOnly selection helpers (See <https://tidyselect.r-lib.org/reference>),  or unquoted expressions representing variable names  can be entered as contents. Changing the names of individual variables using new_name = old_name syntax is allowable",
                             call = caller_call()
                         )
                     }
                 } else if (is.character(x)) {
                     return(as_length_one_quo.character(x))
                 } else {
-                    abort(
+                    cli::cli_abort(
                         "Unexpected entry type in span_structure()",
                         call = caller_call()
                     )
@@ -243,22 +234,15 @@ check_col_plan_dots <- function(x) {
             ) {
                 return(eval_tidy(x))
             } else {
-                stop(
-                    "Invalid entry: `",
-                    format(x),
-                    "`\n",
-                    "Only span_structures (`span_structure()`), ",
-                    "selection helpers (See <https://tidyselect.r-lib.org/reference>), ",
-                    " or unquoted expressions representing variable names ",
-                    " can be entered as contents.",
-                    " Changing the names of individual variables using new_name = old_name syntax is allowable",
-                    call. = FALSE
+                cli::cli_abort(
+                    "Invalid entry: `{format(x)}`\nOnly span_structures (`span_structure()`), selection helpers (See <https://tidyselect.r-lib.org/reference>),  or unquoted expressions representing variable names  can be entered as contents. Changing the names of individual variables using new_name = old_name syntax is allowable",
+                    call = NULL
                 )
             }
         } else if (is.character(x)) {
             return(as_length_one_quo.character(x))
         } else {
-            stop("Unexpected entry type in span_structure()")
+            cli::cli_abort("Unexpected entry type in span_structure()")
         }
     })
 }