+++ tests/testthat/test-tfmt.R
@@ -144,15 +144,7 @@ test_that("basic tfrmt - bare/char mix error", {
             value = value,
             column = column
         ),
-        paste0(
-            "Entries for `",
-            "group",
-            "` argument must be vars(), a character vector, or unquoted column name.\n",
-            "  Consider updating the argument input to `",
-            "group",
-            "` to:\n\t",
-            "vars(row_label1,row_label4)"
-        ),
+        "Entries for `group` argument must be vars(), a character vector, or unquoted column name.  Consider updating the argument input to `group` to: vars(row_label1,row_label4)",
         fixed = TRUE
     )
 })
@@ -431,7 +423,7 @@ test_that("basic tfrmt - ... args", {
     )
     expect_identical(
         message_res,
-        "Argument 'colmn' passed to tfrmt is not a recognized argument.\nDid you intend to use the argument `column`?"
+        "Argument 'colmn' passed to tfrmt is not a recognized argument. Did you intend to use the argument `column`?"
     )
 
     message_res <- capture_messages(
@@ -443,8 +435,8 @@ test_that("basic tfrmt - ... args", {
     expect_identical(
         message_res,
         c(
-            "Argument 'colmn' passed to tfrmt is not a recognized argument.\nDid you intend to use the argument `column`?",
-            "Argument 'lalbl' passed to tfrmt is not a recognized argument.\nDid you intend to use the argument `label`?"
+            "Argument 'colmn' passed to tfrmt is not a recognized argument. Did you intend to use the argument `column`?",
+            "Argument 'lalbl' passed to tfrmt is not a recognized argument. Did you intend to use the argument `label`?"
         )
     )
 })
@@ -461,7 +453,7 @@ test_that("basic tfrmt - erroring args", {
             ) # nolint: missing_argument_linter. Trailing comma is intentional to trigger a missing argument error below
         ),
         paste0(
-            "Error in evaluating argument `body_plan`:\n ",
+            "Error in evaluating argument `body_plan`: ",
             "Error in body_plan(frmt_structure(group_val = \".default\",",
             " label_val = \".default\", : argument is missing, with no default"
         ),
@@ -754,11 +746,7 @@ test_that("basic tfrmt - error when body_plan groups does not match group arg", 
                 )
             )
         ),
-        paste0(
-            "Inconsistencies between group and body_plan\n",
-            "Invalid Format Structure in body_plan at position `1`:\n",
-            "  Malformed Group: invalid, invalid2"
-        ),
+        "Inconsistencies between group and body_plan Invalid Format Structure in body_plan at position `1`: Malformed Group: invalid, invalid2",
         fixed = TRUE
     )
 })
@@ -806,11 +794,7 @@ test_that("basic tfrmt - error when row_grp_plan groups does not match group arg
                 )
             )
         ),
-        paste0(
-            "Inconsistencies between group and row_grp_plan row_grp_structures\n",
-            "Invalid Format Structure in row_grp_plan for row_grp_structure `1`:\n",
-            "  Malformed Group: invalid, invalid2"
-        ),
+        "Inconsistencies between group and row_grp_plan row_grp_structures Invalid Format Structure in row_grp_plan for row_grp_structure `1`: Malformed Group: invalid, invalid2",
         fixed = TRUE
     )
 })
@@ -866,11 +850,7 @@ test_that("basic tfrmt - error when footnote_plan groups does not match group ar
                 marks = "standard"
             )
         ),
-        paste0(
-            "Inconsistencies between group and footnote_plan footnote_structures\n",
-            "Invalid Format Structure in footnote_plan for footnote_structure `2`:\n",
-            "  Malformed Group: invalid, invalid2"
-        ),
+        "Inconsistencies between group and footnote_plan footnote_structures Invalid Format Structure in footnote_plan for footnote_structure `2`: Malformed Group: invalid, invalid2",
         fixed = TRUE
     )
 })
@@ -949,15 +929,9 @@ test_that("basic tfrmt - error when body_plan, row_grp_plan, or footnote_plan gr
             )
         ),
         paste0(
-            "Inconsistencies between group and body_plan\n",
-            "Invalid Format Structure in body_plan at position `1`:\n",
-            "  Malformed Group: invalid, invalid2\n\n",
-            "Inconsistencies between group and row_grp_plan row_grp_structures\n",
-            "Invalid Format Structure in row_grp_plan for row_grp_structure `1`:\n",
-            "  Malformed Group: invalid, invalid2\n\n",
-            "Inconsistencies between group and footnote_plan footnote_structures\n",
-            "Invalid Format Structure in footnote_plan for footnote_structure `1`:\n",
-            "  Malformed Group: invalid, invalid2"
+            "Inconsistencies between group and body_plan Invalid Format Structure in body_plan at position `1`: Malformed Group: invalid, invalid2\n\n",
+            "Inconsistencies between group and row_grp_plan row_grp_structures Invalid Format Structure in row_grp_plan for row_grp_structure `1`: Malformed Group: invalid, invalid2\n\n",
+            "Inconsistencies between group and footnote_plan footnote_structures Invalid Format Structure in footnote_plan for footnote_structure `1`: Malformed Group: invalid, invalid2"
         ),
         fixed = TRUE
     )